+++ crates/fuel-core/src/graphql_api/database.rs
@@ -186,15 +186,11 @@ impl DatabaseMessageProof for ReadView {
     }
 }
 
-impl OnChainDatabase for ReadView {
-    fn owned_message_ids(
-        &self,
-        owner: &Address,
-        start_message_id: Option<Nonce>,
-        direction: IterDirection,
-    ) -> BoxedIter<'_, StorageResult<Nonce>> {
-        self.on_chain
-            .owned_message_ids(owner, start_message_id, direction)
+impl OnChainDatabase for ReadView {}
+
+impl OffChainDatabase for ReadView {
+    fn tx_status(&self, tx_id: &TxId) -> StorageResult<TransactionStatus> {
+        self.off_chain.tx_status(tx_id)
     }
 
     fn owned_coins_ids(
@@ -203,13 +199,17 @@ impl OnChainDatabase for ReadView {
         start_coin: Option<UtxoId>,
         direction: IterDirection,
     ) -> BoxedIter<'_, StorageResult<UtxoId>> {
-        self.on_chain.owned_coins_ids(owner, start_coin, direction)
+        self.off_chain.owned_coins_ids(owner, start_coin, direction)
     }
-}
 
-impl OffChainDatabase for ReadView {
-    fn tx_status(&self, tx_id: &TxId) -> StorageResult<TransactionStatus> {
-        self.off_chain.tx_status(tx_id)
+    fn owned_message_ids(
+        &self,
+        owner: &Address,
+        start_message_id: Option<Nonce>,
+        direction: IterDirection,
+    ) -> BoxedIter<'_, StorageResult<Nonce>> {
+        self.off_chain
+            .owned_message_ids(owner, start_message_id, direction)
     }
 
     fn owned_transactions_ids(